Change log: 1. Added api blueprint and registed to the app. 2. Splitted auth module routes and APIs.
