+++ app/api/__init__.py
@@ -1,0 +1,7 @@
+# pylint: skip-file
+
+from flask import Blueprint
+
+api_bp = Blueprint("api", __name__, url_prefix="/api/v1")
+
+from . import routes
