+++ app/auth/static/js/auth.js
@@ -13,7 +13,7 @@ async function checkEmail(email) {
     console.log("Invalid email");
   }
 
-  const response = await fetch("/v1/check_email_exists/" + email);
+  const response = await fetch("/api/v1/check_email_exists/" + email);
   const res = await response.json();
   console.log(res);
 }
